refactor: change ErrorBoundary
