+++ src/components/SearchBar/SearchBar.tsx
@@ -1,11 +1,16 @@
-import React, { type FC } from 'react';
+import React, { type FC, type FormEventHandler } from 'react';
 import InputField from '../common/InputField/InputField';
 import Button from '../common/Button/Button';
 import type { SearchBarProps } from '../../types/types';
 import './SearchBar.scss';
 
 const SearchBar: FC<SearchBarProps> = (props) => {
-  const { searchValue, searchInputHandler, searchSubmitHandler } = props;
+  const { searchValue, setSearchValue, setShouldUpdateData } = props;
+
+  const searchSubmitHandler: FormEventHandler<HTMLFormElement> = (e) => {
+    e.preventDefault();
+    setShouldUpdateData(true);
+  };
 
   return (
     <div className={'search-bar__wrapper'}>
@@ -14,10 +19,7 @@ const SearchBar: FC<SearchBarProps> = (props) => {
         onSubmit={searchSubmitHandler}
         name={'search-field-form'}
       >
-        <InputField
-          searchValue={searchValue}
-          searchInputHandler={searchInputHandler}
-        />
+        <InputField searchValue={searchValue} setSearchValue={setSearchValue} />
         <Button title={'Search'} classType={'submit-button'} />
       </form>
     </div>
