+++ src/pages/Main/Main.tsx
@@ -1,10 +1,10 @@
-import React, { FC, useEffect, useRef, useState } from 'react';
+import React, { type FC, useEffect, useRef, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { fetchPeople } from '../../api/fetchPeople';
 import NavBar from '../../components/NavBar/NavBar';
 import PersonDetails from '../../components/PersonDetails/PersonDetails';
 import type { MainProps } from '../../types/types';
-import { Person } from '../../types/apiTypes';
+import type { Person } from '../../types/apiTypes';
 
 const Main: FC<MainProps> = (props) => {
   const {
